P16-351: Fixing compiler warnings
We removed some comments that didn't add value. This included some fall through
statements in a switch. We needed those to remove compiler warnings about
implicit fall through.

Also, fixed one in legacy.cpp that was still hanging around.

+++ sdk/src/pixie/pixie16/legacy.cpp
@@ -77,7 +77,7 @@ void settings::load(const std::string& parfile) {
     }
 
     dsp_mem.clear();
-    dsp_mem.resize(size / sizeof(hw::word));
+    dsp_mem.resize(static_cast<unsigned int>(size) / sizeof(hw::word));
     input.read(reinterpret_cast<char*>(dsp_mem.data()), size);
 }
